requirements.txt
aiohttp==3.8.6
PyNaCl==1.5.0
Requests==2.32.3
